requirements.txt
pypdf 
langchain-google-genai 
langchain_community 
chromadb
google-generativeai
python-dotenv